pyproject.toml
[project]
name = "dio-bank"
version = "0.1.0"
description = "Portfolio-ready Flask blog API with JWT authentication, role-based access control, SQLAlchemy, and automated tests."
authors = [
    {name = "uilliambruno-droid",email = "uilliambruno@gmail.com"}
]
requires-python = ">=3.14,<4"
dependencies = [
    "flask (>=3.1.3,<4.0.0)",
    "flask-sqlalchemy",
    "flask-migrate",
    "flask-jwt-extended (>=4.7.1,<5.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "gunicorn (>=25.1.0,<26.0.0)",
    "flask-bcrypt (>=1.0.1,<2.0.0)",
    "flask-marshmallow (>=1.4.0,<2.0.0)",
    "marshmallow-sqlalchemy (>=1.4.2,<2.0.0)",
    "apispec-webframeworks (>=1.2.0,<2.0.0)",
]

# pytest.toml or .pytest.toml
[pytest]
addopts = ["-ra", "-q"]
testpaths = [
    "tests"
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "pytest (>=9.0.2,<10.0.0)",
    "pytest-mock (>=3.15.1,<4.0.0)"
]
